requirements.txt
PyQt5>=5.15.0
requests>=2.25.1
python-dotenv>=0.19.0
matplotlib>=3.4.0
numpy>=1.19.0
urllib3<2.0.0
pytest==7.3.1 